error: remove error.rs and move all its content to lib.rs
